# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Some endpoints for the tem db",
        "title": "TEM Backend",
        "contact": {
            "name": "Flou21",
            "email": "muehlhans.f@coflnet.com"
        },
        "license": {
            "name": "AGPL v3"
        },
        "version": "1.0"
    },
    "host": "sky.coflnet.com/tem/",
    "basePath": "/api/",
    "paths": {
        "/player/{uuid}": {
            "get": {
                "description": "get a player by his player uuid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "player"
                ],
                "summary": "PlayerUUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github.com_Coflnet_tem-backend_internal_api.PlayerResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github.com_Coflnet_tem-backend_internal_api.PlayerResponse": {
            "type": "object",
            "properties": {
                "generic_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mongo.GenericItem"
                    }
                },
                "generic_pets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mongo.GenericPet"
                    }
                },
                "id": {
                    "$ref": "#/definitions/mongo.PlayerId"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mongo.Item"
                    }
                }
            }
        },
        "internal_api.PlayerResponse": {
            "type": "object",
            "properties": {
                "generic_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mongo.GenericItem"
                    }
                },
                "generic_pets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mongo.GenericPet"
                    }
                },
                "id": {
                    "$ref": "#/definitions/mongo.PlayerId"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mongo.Item"
                    }
                }
            }
        },
        "mongo.GenericItem": {
            "type": "object",
            "properties": {
                "enchantments": {},
                "extra_attributes": {},
                "id": {
                    "type": "string"
                },
                "item_id": {
                    "type": "string"
                },
                "rarity": {
                    "type": "string"
                },
                "reforge": {
                    "type": "string"
                }
            }
        },
        "mongo.GenericPet": {
            "type": "object",
            "properties": {
                "candy": {
                    "type": "integer"
                },
                "held_item": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "rarity": {
                    "type": "string"
                },
                "skin": {
                    "type": "string"
                }
            }
        },
        "mongo.Item": {
            "type": "object"
        },
        "mongo.PlayerId": {
            "type": "object"
        }
    }
}